Set index, newindex metamethods on _ENV table without creating global variables
